18119 nvme_print_identify_ctrl_list() needs to verify id count 18120 nvmeadm phy eye logic has mult overflow bugs Reviewed by: Andy Fiddaman <illumos@fiddaman.net> Reviewed by: Dan McDonald <danmcd@edgecast.io> Approved by: Gordon Ross <gordon.w.ross@gmail.com>

+++ usr/src/common/nvme/nvme_log.c
@@ -323,7 +323,7 @@ nvme_lpd_eom_len(uint64_t *outp, const void *data, size_t len)
 	case NVME_EOM_IN_PROGRESS:
 		break;
 	case NVME_EOM_DONE:
-		out += hdr.eom_ds * hdr.eom_nd;
+		out += (uint64_t)hdr.eom_ds * (uint64_t)hdr.eom_nd;
 		break;
 	default:
 		/*
